+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
                   android:label="Photo gallery"/>
 
         <meta-data android:name="AA_DB_NAME" android:value="Vkclient.db"/>
-        <meta-data android:name="AA_DB_VERSION" android:value="2"/>
+        <meta-data android:name="AA_DB_VERSION" android:value="3"/>
         <meta-data
             android:name="AA_MODELS"
             android:value="com.terebenin.vkclient.models.newsItem.Item,
@@ -32,7 +32,7 @@
                            com.terebenin.vkclient.models.newsItem.Profile"/>
         <meta-data
             android:name="AA_SERIALIZERS"
-            android:value="com.terebenin.vkclient.models.newsItem.AttachmentSerializer "/>
+            android:value="com.terebenin.vkclient.models.newsItem.AttachmentListSerializer "/>
     </application>
 
 </manifest>
