--- app/src/main/java/com/terebenin/vkclient/models/newsItem/AttachmentSerializer.java
@@ -1,35 +1,0 @@
-package com.terebenin.vkclient.models.newsItem;
-
-import com.activeandroid.serializer.TypeSerializer;
-
-/**
- * Created by evgeny on 24.02.17.
- */
-
-public class AttachmentSerializer extends TypeSerializer {
-    @Override
-    public Class<?> getDeserializedType() {
-        return Attachment.class;
-    }
-
-    @Override
-    public Class<?> getSerializedType() {
-        return String.class;
-    }
-
-    @Override
-    public String serialize(Object data) {
-        if (data == null) {
-            return null;
-        }
-        return ((Attachment) data).getType();
-    }
-
-    @Override
-    public Attachment deserialize(Object data) {
-        if (data == null) {
-            return null;
-        }
-        return new Attachment((String) data);
-    }
-}
